+++ sdk/src/public/java/com/mapbox/maps/MapboxMap.kt
@@ -1,6 +1,7 @@
 package com.mapbox.maps
 
 import android.app.Activity
+import android.graphics.RectF
 import android.os.Handler
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
@@ -863,6 +864,46 @@ class MapboxMap :
   override fun coordinatesForPixels(pixels: List<ScreenCoordinate>): List<Point> {
     checkNativeMap("coordinatesForPixels")
     return nativeMap.coordinatesForPixels(pixels)
+  }
+
+  /**
+   * Calculates geographical `coordinates` (i.e., longitude-latitude pairs) that correspond
+   * to a [RectF] that holds four screen points.
+   *
+   * The screen points are in `platform pixels` relative to the top left corner
+   * of the map (not of the whole screen).
+   *
+   * This API isn't supported by Globe projection.
+   *
+   * @param rectF Rectangle with 4 edges (left, top, right, bottom).
+   *
+   * @return List of `geographical coordinates` that corresponds to given edges of RectF
+   * in clockwise direction starting from topLeft. if provided [RectF] is empty, an
+   * [IllegalArgumentException] will be thrown.
+   */
+  override fun coordinatesForRect(rectF: RectF): List<Point> {
+    checkNativeMap("coordinatesForRect")
+    if (rectF.isEmpty) throw IllegalArgumentException("RectF must not be empty")
+    val screenCoordinateTopLeft = ScreenCoordinate(
+      rectF.left.toDouble(), rectF.top.toDouble()
+    )
+    val screenCoordinateTopRight = ScreenCoordinate(
+      rectF.right.toDouble(), rectF.top.toDouble()
+    )
+    val screenCoordinateBottomRight = ScreenCoordinate(
+      rectF.right.toDouble(), rectF.bottom.toDouble()
+    )
+    val screenCoordinateBottomLeft = ScreenCoordinate(
+      rectF.left.toDouble(), rectF.bottom.toDouble()
+    )
+    return nativeMap.coordinatesForPixels(
+      listOf(
+        screenCoordinateTopLeft,
+        screenCoordinateTopRight,
+        screenCoordinateBottomRight,
+        screenCoordinateBottomLeft
+      )
+    )
   }
 
   /**