+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Make padding optional for `MapboxMap.cameraForCoordinateBounds`, `MapboxMap.cameraForCoordinates`, `MapboxMap.cameraForGeometry` methods.
 * Add `FreeCameraOptions.getLocation` and `FreeCameraOptions.getAltitude` methods.
 * Remove `MapAnimationOptions.animatorListener` property. In order to subscribe to animations, provide `Animator.animatorListener` with `flyTo`, `easeTo`, `pitchBy`, `scaleBy`, `moveBy`, `rotateBy` apis.
+* Add `MapboxMap.coordinatesForRect(rectF: RectF)` to support rectangle parameters.
  
 ## Dependencies
 * Update dependencies